@@ -17,7 +17,7 @@ dotenv.config();
 const app = express();
 
 app.use(cors({
-  origin:'http://localhost:5173',
+  origin: process.env.CLIENT_URL,
   credentials: true,
 }));
 
