@@ -1,8 +1,10 @@
 export const MESSAGES = {
   NOT_FOUND: "Not found",
-  INVALID: "Invalid request",
+  INVALID_REQUEST: "Invalid request",
   FORBIDDEN: "Forbidden",
   CREATED: "Created",
   UPDATED: "Updated",
   DELETED: "Deleted",
+  BANNED: "Banned",
+  OTP_SENT: "OTP sent successfully",
 } as const;
