@@ -7,15 +7,15 @@ import { MESSAGES } from '../constants/messages'
 @injectable()
 export class DietService {
   constructor (
-    @inject(TYPES.IDietDayRepository) private repo: IDietDayRepository
+    @inject(TYPES.IDietDayRepository) private _repo: IDietDayRepository
   ) {}
 
   async createOrGetDay (userId: string, date: string): Promise<IDietDay> {
-    return this.repo.createOrGet(userId, date)
+    return this._repo.createOrGet(userId, date)
   }
 
   async getDay (userId: string, date: string): Promise<IDietDay | null> {
-    return this.repo.getByUserAndDate(userId, date)
+    return this._repo.getByUserAndDate(userId, date)
   }
 
   async addMeal (
@@ -47,8 +47,8 @@ export class DietService {
 
     mealPayload.usedBy = userId
 
-    await this.repo.createOrGet(userId, date)
-    return this.repo.addMeal(userId, date, mealPayload as IMeal)
+    await this._repo.createOrGet(userId, date)
+    return this._repo.addMeal(userId, date, mealPayload as IMeal)
   }
 
   async updateMeal (
@@ -58,7 +58,7 @@ export class DietService {
     mealId: string,
     update: Partial<IMeal>
   ): Promise<IDietDay | null> {
-    const day = await this.repo.getByUserAndDate(userId, date)
+    const day = await this._repo.getByUserAndDate(userId, date)
     if (!day) throw new Error(MESSAGES.NOT_FOUND)
 
     const meal = day.meals.find(m => m._id?.toString() === mealId)
@@ -76,7 +76,7 @@ export class DietService {
         throw new Error(MESSAGES.FORBIDDEN)
     }
 
-    return this.repo.updateMeal(userId, date, mealId, update)
+    return this._repo.updateMeal(userId, date, mealId, update)
   }
 
   async markMealEaten (
@@ -91,7 +91,7 @@ export class DietService {
     if (actor.role !== 'user' && actor.role !== 'trainer')
       throw new Error(MESSAGES.FORBIDDEN)
 
-    return this.repo.markMeal(userId, date, mealId, isEaten)
+    return this._repo.markMeal(userId, date, mealId, isEaten)
   }
 
   async removeMeal (
@@ -100,7 +100,7 @@ export class DietService {
     date: string,
     mealId: string
   ): Promise<IDietDay | null> {
-    const day = await this.repo.getByUserAndDate(userId, date)
+    const day = await this._repo.getByUserAndDate(userId, date)
     if (!day) throw new Error(MESSAGES.NOT_FOUND)
     const meal = day.meals.find(m => m._id?.toString() === mealId)
     if (!meal) throw new Error(MESSAGES.NOT_FOUND)
@@ -118,6 +118,6 @@ export class DietService {
         throw new Error(MESSAGES.FORBIDDEN)
     }
 
-    return this.repo.removeMeal(userId, date, mealId)
+    return this._repo.removeMeal(userId, date, mealId)
   }
 }
