@@ -3,14 +3,14 @@ import { OAuth2Client } from "google-auth-library";
 
 @injectable()
 export class OAuthClient {
-  private client: OAuth2Client;
+  private _client: OAuth2Client;
 
   constructor() {
-    this.client = new OAuth2Client(process.env.GOOGLE_CLIENT_ID);
+    this._client = new OAuth2Client(process.env.GOOGLE_CLIENT_ID);
   }
 
   async verifyIdToken(token: string) {
-    const ticket = await this.client.verifyIdToken({
+    const ticket = await this._client.verifyIdToken({
       idToken: token,
       audience: process.env.GOOGLE_CLIENT_ID,
     });
